DataStructures

Description

Datatypes used in the application.

Written by Johannes Hartmann, Liam Kelly, Manuel Campos Villarreal

Documentation

data Movie

Data structure representig a movie. The first parameter is the ID and the second is the name of the movie and the third the release day

Constructors

Movie Int String String

data Actor

Data structure representig a actor. The first parameter is the ID, the second is the name of the actor and the third are all movies he plays in

Constructors

Actor Int String [Movie]

data Cinema

Data structure representig a cinema. The first parameter is the ID, the second is the name of the cinema and the third is the range to the given location

Constructors

Cinema String String Float

newtype Movie2

Data structure representing a Movie from the second API

Constructors

Movie2 String