HTTPRequestModule

Description

This module provides all methodes neccesary for fillig the datanase with movies and actores

Written by Liam Kelly, Johannes Hartmann

Documentation

httpGetListOfMovies :: String -> IO [Movie]

returns all Movies starting from a given time string

httpGetListOfActores :: [Movie] -> IO [Actor]

returns a List of all actors playing in the given list of movies