JSONParserModule

Description

This module takes care of the json parsing for each http module

Written by Liam Kelly, Manuel Campos Villarreal, Johannes Hartmann

Documentation

parseMovies :: ByteString -> [Movie]

parseActors :: ByteString -> Movie -> [Actor]

parsePages :: ByteString -> Int

parseCinemas :: ByteString -> [Cinema]

parseMovies2 :: ByteString -> [Movie2]

data MovieFromJSON

data Results

Orphan instances

FromJSON Movie2

Methods

parseJSON :: Value -> Parser Movie2

FromJSON Movie

Methods

parseJSON :: Value -> Parser Movie

FromJSON Cinema

Methods

parseJSON :: Value -> Parser Cinema